+++ src/main/java/by/grsu/liceum/service/impl/GroupServiceImpl.java
@@ -6,8 +6,10 @@ import by.grsu.liceum.dto.group.GroupFullDto;
 import by.grsu.liceum.dto.mapper.GroupDtoMapper;
 import by.grsu.liceum.entity.Account;
 import by.grsu.liceum.entity.Group;
+import by.grsu.liceum.entity.Institution;
 import by.grsu.liceum.exception.AccountWithIdNotFoundException;
 import by.grsu.liceum.exception.GroupWithIdNotFoundException;
+import by.grsu.liceum.exception.InstitutionWithIdNotFoundException;
 import by.grsu.liceum.exception.InvalidPermissionsException;
 import by.grsu.liceum.exception.NullableGroupCreationDtoException;
 import by.grsu.liceum.repository.AccountRepository;
@@ -26,25 +28,35 @@ public class GroupServiceImpl implements GroupService {
     private final GroupRepository groupRepository;
 
     @Override
-    public GroupFullDto findGroupById(long id) {
+    public GroupFullDto findGroupById(long institutionId, long id) {
         Group group = Optional.ofNullable(groupRepository.findById(id))
                 .orElseThrow(() -> new GroupWithIdNotFoundException(id));
+
+        if (group.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
 
         return GroupDtoMapper.convertEntityToFullDto(group);
     }
 
     @Override
     @Transactional
-    public GroupFullDto createNewGroup(GroupCreationDto creationDto) {
+    public GroupFullDto createNewGroup(long institutionId, GroupCreationDto creationDto) {
         if(creationDto == null)
             throw new NullableGroupCreationDtoException();
 
         Account admin = Optional.ofNullable(accountRepository.findById(creationDto.getAdminId()))
                 .orElseThrow(() -> new AccountWithIdNotFoundException(creationDto.getAdminId()));
 
+        Institution institution = Optional.of(admin.getInstitution())
+                .orElseThrow(() -> new InstitutionWithIdNotFoundException(institutionId));
+
+        if(admin.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
+
         Group group = Group.builder()
                 .admin(admin)
                 .name(creationDto.getName())
+                .institution(institution)
                 .build();
 
         groupRepository.save(group);
@@ -59,13 +71,18 @@ public class GroupServiceImpl implements GroupService {
             member.getOtherGroups().add(group);
         }
 
+        institution.getGroups().add(group);
+
         return GroupDtoMapper.convertEntityToFullDto(group);
     }
 
     @Override
     @Transactional
-    public void addMembersToTheGroup(AddMembersDto addMembersDto) {
+    public void addMembersToTheGroup(long institutionId, AddMembersDto addMembersDto) {
         Group group = groupRepository.findById(addMembersDto.getGroupId());
+
+        if(group.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
 
         if(group.getAdmin().getId() != addMembersDto.getAdminId())
             throw new InvalidPermissionsException();
@@ -80,8 +97,12 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void deleteGroupById(long id) {
-        findGroupById(id);
+    public void deleteGroupById(long institutionId, long id) {
+        Group group = Optional.ofNullable(groupRepository.findById(id))
+            .orElseThrow(() -> new GroupWithIdNotFoundException(id));
+
+        if(group.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
 
         groupRepository.deleteById(id);
     }
